# apps/dashboard/lib/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Carbon API",
    "description": "Control plane for Carbon — projects, ingestion, emulators, snapshots, artifacts, org/billing, and CLI auth. Every endpoint accepts either an API key (`x-carbon-key`) or a Better Auth session cookie/bearer token; the docs and `/health` endpoints are the only routes served anonymously.",
    "version": "dev",
    "contact": {
      "name": "Carbon",
      "url": "https://github.com/TaxCollector23/carbon",
      "email": "support@example.com"
    },
    "license": { "name": "Apache-2.0", "url": "https://www.apache.org/licenses/LICENSE-2.0" }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-carbon-key",
        "description": "A key minted via `POST /v1/api-keys` or the `bootstrap` script. Prefix `ck_live_*` for production, `ck_test_*` for test envs."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "better-auth-session",
        "description": "A Better Auth session bearer token. Preferred for programmatic access from a signed-in user."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "The Better Auth session cookie set by the dashboard. Used implicitly when the browser calls the API cross-origin with `credentials: \"include\"`."
      }
    },
...
  },
  "paths": {
...
    "/v1/ingest": {
      "post": {
        "summary": "Ingest a spec into a project",
        "tags": ["Ingest"],
        "description": "Turn an OpenAPI / HAR / Postman payload into an IR and behavior graph. Set `async: true` to enqueue the work and return a `jobId` immediately (recommended for large specs); the sync path returns the persisted artifact ids and counts.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectSlug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[a-z0-9][a-z0-9-]*$"
                  },
                  "source": {
                    "anyOf": [
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["json"] },
                          "content": {},
                          "hint": { "type": "string" }
                        },
                        "required": ["kind"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["text"] },
                          "content": { "type": "string" },
                          "hint": { "type": "string" }
                        },
                        "required": ["kind", "content"],
                        "additionalProperties": false
                      }
                    ]
                  },
                  "origin": { "type": "string" },
                  "enrich": { "type": "boolean", "default": false },
                  "async": { "type": "boolean", "default": false }
                },
                "required": ["projectSlug", "source"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "irId": { "type": "string" },
                    "graphId": { "type": "string" },
                    "api": {},
                    "endpoints": { "type": "integer" },
                    "resources": { "type": "integer" },
                    "warnings": { "type": "array", "items": {} },
                    "judge": {}
                  },
                  "required": ["irId", "graphId", "endpoints", "resources"],
                  "additionalProperties": true
                }
              }
            }
          },
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "jobId": { "type": "string" }, "status": { "type": "string" } },
                  "required": ["jobId", "status"],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/postman": {
      "post": {
        "summary": "Ingest a raw Postman collection",
        "tags": ["Ingest"],
        "description": "Shortcut that accepts a raw Postman collection JSON body (any content) and routes it through the Postman adapter. Target project is picked via `?projectSlug=`.",
        "parameters": [
          {
            "schema": { "type": "string" },
            "in": "query",
            "name": "projectSlug",
            "required": true
          },
          { "schema": { "type": "string" }, "in": "query", "name": "origin", "required": false }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "irId": { "type": "string" },
                    "graphId": { "type": "string" },
                    "api": {},
                    "endpoints": { "type": "integer" },
                    "resources": { "type": "integer" },
                    "warnings": { "type": "array", "items": {} },
                    "judge": {}
                  },
                  "required": ["irId", "graphId", "endpoints", "resources"],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [{ "url": "http://localhost:4000", "description": "local dev" }],
  "security": [{ "apiKey": [] }, { "bearerAuth": [] }, { "sessionCookie": [] }],
...
}